Split the check-in test into rendering, flows, and routing suites
